Add Start app data context
